+++ src/loading/scene/mod.rs
@@ -1,5 +1,7 @@
 mod scene_builder;
+mod scene_handle_error;
 mod spawn_scene_ext;
 
 pub use scene_builder::*;
+pub use scene_handle_error::*;
 pub use spawn_scene_ext::*;