+++ src/sickle_ext/interaction_ext.rs
@@ -82,7 +82,7 @@ pub trait UiInteractionExt
     /// Adds a reactor to a [`PointerEnter`] entity event.
     ///
     /// Equivalent to `entity_builder.on_event::<PointerEnter>().r(callback)`.
-    fn on_pointer_enter<R: ReactorResult, M>(
+    fn on_pointer_enter<R: CobwebResult, M>(
         &mut self,
         callback: impl IntoSystem<(), R, M> + Send + Sync + 'static,
     ) -> &mut Self;
@@ -90,7 +90,7 @@ pub trait UiInteractionExt
     /// Adds a reactor to a [`PointerLeave`] entity event.
     ///
     /// Equivalent to `entity_builder.on_event::<PointerLeave>().r(callback)`.
-    fn on_pointer_leave<R: ReactorResult, M>(
+    fn on_pointer_leave<R: CobwebResult, M>(
         &mut self,
         callback: impl IntoSystem<(), R, M> + Send + Sync + 'static,
     ) -> &mut Self;
@@ -98,7 +98,7 @@ pub trait UiInteractionExt
     /// Adds a reactor to a [`Pressed`] entity event.
     ///
     /// Equivalent to `entity_builder.on_event::<Pressed>().r(callback)`.
-    fn on_pressed<R: ReactorResult, M>(
+    fn on_pressed<R: CobwebResult, M>(
         &mut self,
         callback: impl IntoSystem<(), R, M> + Send + Sync + 'static,
     ) -> &mut Self;
@@ -106,7 +106,7 @@ pub trait UiInteractionExt
     /// Adds a reactor to a [`Released`] entity event.
     ///
     /// Equivalent to `entity_builder.on_event::<Released>().r(callback)`.
-    fn on_released<R: ReactorResult, M>(
+    fn on_released<R: CobwebResult, M>(
         &mut self,
         callback: impl IntoSystem<(), R, M> + Send + Sync + 'static,
     ) -> &mut Self;
@@ -114,7 +114,7 @@ pub trait UiInteractionExt
     /// Adds a reactor to a [`PressCanceled`] entity event.
     ///
     /// Equivalent to `entity_builder.on_event::<PressCanceled>().r(callback)`.
-    fn on_press_canceled<R: ReactorResult, M>(
+    fn on_press_canceled<R: CobwebResult, M>(
         &mut self,
         callback: impl IntoSystem<(), R, M> + Send + Sync + 'static,
     ) -> &mut Self;
@@ -122,7 +122,7 @@ pub trait UiInteractionExt
 
 impl UiInteractionExt for UiBuilder<'_, Entity>
 {
-    fn on_pointer_enter<R: ReactorResult, M>(
+    fn on_pointer_enter<R: CobwebResult, M>(
         &mut self,
         callback: impl IntoSystem<(), R, M> + Send + Sync + 'static,
     ) -> &mut Self
@@ -132,7 +132,7 @@ impl UiInteractionExt for UiBuilder<'_, Entity>
         self
     }
 
-    fn on_pointer_leave<R: ReactorResult, M>(
+    fn on_pointer_leave<R: CobwebResult, M>(
         &mut self,
         callback: impl IntoSystem<(), R, M> + Send + Sync + 'static,
     ) -> &mut Self
@@ -142,7 +142,7 @@ impl UiInteractionExt for UiBuilder<'_, Entity>
         self
     }
 
-    fn on_pressed<R: ReactorResult, M>(
+    fn on_pressed<R: CobwebResult, M>(
         &mut self,
         callback: impl IntoSystem<(), R, M> + Send + Sync + 'static,
     ) -> &mut Self
@@ -152,7 +152,7 @@ impl UiInteractionExt for UiBuilder<'_, Entity>
         self
     }
 
-    fn on_released<R: ReactorResult, M>(
+    fn on_released<R: CobwebResult, M>(
         &mut self,
         callback: impl IntoSystem<(), R, M> + Send + Sync + 'static,
     ) -> &mut Self
@@ -162,7 +162,7 @@ impl UiInteractionExt for UiBuilder<'_, Entity>
         self
     }
 
-    fn on_press_canceled<R: ReactorResult, M>(
+    fn on_press_canceled<R: CobwebResult, M>(
         &mut self,
         callback: impl IntoSystem<(), R, M> + Send + Sync + 'static,
     ) -> &mut Self