+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### Highlights
+- `SceneHandle` callbacks return `impl CobwebResult`.
+- Update `SceneHandle` methods to return `Result<x, SceneHandleError`. This makes the API panic-free.
+
+### Updates
+- Update to `bevy_cobweb` 0.16. This comes with better error messages.
+
 ## 0.10.0
 
 ### Highlights